PQLParser::eat_*_string - Take in const char * instead of string
Using the newly added StringBuffer::strcmp, we are now able to compare
the StringBuffer's contents with const char * directly. This allows us
to replace the eat_*_string methods' second parameter with their original
type - const char *. This helps us avoid the overhead of casting the
const char * (most of which are #defined) up to std::string and avoid
converting the StringBuffer's contents into std::string.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>