PQLParser - introduce eat_relRef_generic
All RelRefs are of the following form:

    RelRefName '(' arg1 ',' arg2 ')'

They have:
- A name [and a PQLParser method to eat that exact name], optionally
  followed by whitespace
- A left parentheses, optionally followed by whitespace
- A first argument, optionally followed by whitespace
- A comma, optionally followed by whitespace
- A second argument, optionally followed by whitespace
- A right parentheses

These steps are very similar. In particular, since we have methods to
match the RelRefName, left paren, comma, right paren, we can pass
such methods to a generic function which does all these automatically.

That leaves the 2 arguments. They can be of the following types:

    entRef, varRef
    entRef, entRef
    stmtRef, varRef
    stmtRef, stmtRef
    lineRef, lineRef

In particular, in terms of lexemes:

    lineRef = stmtRef
    stmtRef < entRef
    varRef < entRef

And we can implement methods to parse different combinations of such
arguments. Currently, this is done for entRef,varRef in
PQLParser::eat_entRef_varRef.

An instantiation of eat_relRef_generic for different RelRef simply
requires a change in the methods passed in. In particular:
- The eat_* method (eat_modifies, eat_calls_star, etc) used to match
  the RelRefName
- The eat_XRef_YRef method used to parse the parameters

A concrete use of eat_relRef_generic is in PQLParser::eat_relRef_modifies

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>