PQLParser::error - throw exception
When the error method is called, there is an unrecoverable error with
the parse. We adopt the approach of erroring out as soon as we discover
the first error. As such, it does seem appropriate to halt the parse
at that point and avoid wasting time trying other productions of the
PQL grammar since they will fail anyway.

We introduce a ParseError enumeration which contains the various types
of parsing errors.

When PQLParser::error finishes printing the error messages, it will
throw an exception which is caught at PQLParser::parse.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>