StringBuffer - add c_str method
This allows us to directly use the StringBuffer's internal character
array for reading and avoids the cost of converting the buffer to a
C++ string, then calling its c_str method.

However, one has to take special caution not to append a StringBuffer's
c_str return value to itself, as in:

    StringBuffer sb;
    // do some stuff here
    sb.append(sb.c_str());

Because this will result in an infinite loop. A comment is present
in StringBuffer.h to remind users of this.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>