PQLParser - Revamp error system
Whenever PQLParser::error is invoked directly or indirectly, there will
be an appropriate error string with placeholder arguments ("%d", "%s")
available at PQLParser::print_error. This centralizes the location of
all error strings for all parsing errors and allows callers to substitute
actual arguments in place of the placeholders, thereby standardizing the
errror messages. This is made possible by the StringBuffer::sprintf and
StringBuffer::vsprintf methods introduced in the previous patches.

Several helper functions have been introduced to tailor to the requirements
of error messages for different errors. They are:

- error_add_select_tuple
  -> for Select construct, eg. Select <a,b,c>
- error_set_relRef_arg
  -> for errors encountered during setting of RelRef arguments
- error_add_relRef
  -> for errors encountered during finalizing of RelRef arguments.
     This is currently for undeclared synonyms and type errors.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>