PQLParser::eat_decl_one - Eat till comma/space/semicolon on error
PQLParser::eat_synonym halts as soon as the first character on the
input buffer is not a letter. Therefore, when it returns false, it makes
no sense to check whether the string consumed is empty since it either
consumes a string and returns true or does not consume any character
and returns false.

We introduce the method PQLParser::eat_till_comma_space_semicolon
which consumes characters from the input buffer until we encounter
a comma, space or semicolon. This allows us to consume as many characters
as possible and disambiguate between a non-empty synonym vs. an incorrect
synonym.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>