Introduce PQLParser::eat_design_entity and deprecate eat_nonws_token
PQLParser::eat_design_entity eats one the 10 design entities defined
in the PQL grammar. 9 out of 10 of these design entities are alphabetic
strings while the other one is a string that consists of alphabets and
underscore (prog_line). As such, we make use of the defined
eat_alpha_string method for parsing the 9 alphabetic design entities
and we introduce the eat_alpha_underscore and
eat_alpha_underscore_string methods for parsing prog_line.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>