PQLParser - Rename PARSE_SELECT_UNKNOWN to PARSE_SELECT_UNDECLARED and add test case
This better reflects the nature of this error. An example:

    stmt s;
    Select a // a is undeclared

We error out in PQLParser::eat_select_tuple_elem whenever an undeclared
synonym is parsed since that is the only possible parse.

In PQLParser::_parse, we reordered the parsing order to try to parse
'Select BOOLEAN' first. Otherwise, we will try to parse tuples and
the 'BOOLEAN' will be incorrectly parsed as a synonym and result in
PARSE_SELECT_UNDECLARED error being issued.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>