PQLParser - Introduce PARSE_SELECT_UNDEF_ATTRNAME error
This error is thrown for undefined attribute names after we have
successfully parsed a synonym and a period. Because an attrRef is the
only place where a period can appear in a select tuple, it is perfectly
valid to error out here and stop parsing.

We only have four valid attribute names, namely "procName", "varName",
"value" and "stmt#".

Examples of such errors are tested in
TestPQLParser::test_err_select_undef_attrname in this patch.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>