StringBuffer::grow_buffer - grow to at least current chars + required chars
Previously, whenever we encountered a situation which requires us to grow
our buffer, we would expand it by 1.5 times. However, this may not be
sufficient for the new capacity requirements.

Now, we correct StringBuffer::grow_buffer to grow to at least what is
required of it to hold the additional characters.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>