string_to_uint - Return false upon discovering string exceeds 10 chars
Otherwise, we will carry on to actually converting the integer, which
is totally unintended since it may overflow the long long datatype and
return a wrong answer.

Furthermore, in the case wheere it fits within a long long, the error
message will be overwritten after integer converson and become
S_TO_UINT_OVERFLOW instead of S_TO_UINT_TOO_LONG. This not only
results in a memory leak but also the wrong message being returned.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>