PQLParser - remove PARSE_REL_ARG_INVALID
This error is only triggered when an argType of RELARG_INVALID is passed
into RelRef::set_arg_* . The caller to RelRef::set_arg_* should have done
a check prior to calling the method. As such, this error is unnecessary
and hence removed.

An assertion is added in RelRef::set_arg to ensure that
argType != RELARG_INVALID

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>