PQLParser - abstract eat_entRef_varRef into eat_XRef_YRef
Most of the code for eating the 2 arguments of any relRef can be reused,
except for these 3 parts:

- Eating the relRef name
- Eating the first argument
- Eating the second argument

We have identified this in PQLParser::eat_relRef_generic, and
PQLParser::eat_entRef_varRef does the eating of the first and second
arguments of Modifies. However, most of the eat_entRef_varRef code
applies to any type of Ref, not just entRef and varRef. As such,
we slightly modify it into a method which accepts 2 methods for
each of the 2 arguments involved in a relRef. The result is in
PQLParser::eat_XRef_YRef.