Rename test files to reorder the order of execution
Our test files were previously named:
TestPQLParser
TestStringBuffer

As a result, the TestPQLParser test cases were always executed before
the ones for StringBuffer. However, PQLParser depends on StringBuffer.
As such, we rename the test files so as to ensure that test cases for
StringBuffer execute before test cases for PQLParser.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>