QueryInfo - add insert order for clauses
Previously, we used the evalOrder vector of QueryInfo to record
the order of clauses in a PQL Query.

However, the actual usage of evalOrder is to store the actual evaluation
order after we process QueryInfo with a Query Optimizer.

Therefore, we create an insertOrder vector to maintain the insert order
of PQL clauses.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>