QueryInfo - Introduce optimize method
Currently, the optimize method does not actually optimize anything.
It simply copies the insertOrder vector to the evalOrder vector.

We also add a dump_optimized_to_string method to dump the optimized
clauses. This will mean that the select clauses will be empty if
the optimize method was not called.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>