QueryEvaluator - Scaffolding code up
This is a rather huge patch which incorporates a number of changes,
mainly to the QueryEvaluator and PKB.

For PKB, we introduce a few new methods:
- get_all_vars_modified_by_assign
- get_all_assign_modifying_var
- get_all-assign

For VarTable, we separate modifies into several different categories:
- assign statements that modify the given variable
- if statements that modify the given variable
- while statements that modify the given variable
- call statements that modify the given variable
- all statements that modify the given variable

This allows us to easily distinguish the different types of statements
that modify a variable. Instead of doing the work in the QueryEvaluator
as we did in the past, we now shift the work to the PKB.

For the QueryEvaluator, it is observed that for RelRef evaluation
(Modifies, Uses, Follows, Parent, etc), a lot of code will likely
be very similar. The differences in code will boil down to:

- Type of RelRef (whether it is Modifies, Uses, Affects, Parent, etc)
- Type of 1st argument of RelRef
- Type of 2nd argument of RelRef
- Whether 1st argument is synonym, is it in the results graph
- Whether 2nd argument is synonym, is it in the results graph
- If 1st argument is not a synonym, what type is it
- If 2nd argument is not a synonym, what type is it

Hence, the EvalSynArgDesc is created, encompassing the information
above. This allows us to dispatch to a proper function which will
thereby process the RelRef and avoid a lot of code duplication.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>