ResultsProjector - add all of synonym in if it's not in graph
Previously when generating the results, we regarded the whole query
as false if a particular synonym inside the select clause is not inside
the ResultsGraph.

However, this is a mistake. In actuality, a synonym which is not inside
the ResultsGraph is one which is not involved in any clause. Therefore,
instead of regarding the query as evaluating to false, we should include
all synonyms of that particular type into the ResultsGraph.

This bug is caught by our test case in Test_00_SelectNoClause

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>