PKB - store all constants as integers, add get_all_const method
Previously, constants in SIMPLE were stored as strings. However,
this is slightly inefficient whenever we need to compare / search
for a constant and it assumes that we can handle arbitrary precision
integer constants.

Now, we set the upper bound of an integer constant to 2^31-1, or
2147483647. This allows us to use the C++ int datatype to store integer
constants and leads to slightly more efficient lookup operations.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>