PKB - add get_all_stmtLst method
This allows us to retrieve all stmtLst in a given SIMPLE program.
A stmtLst number is defined as the statement number of the
first statement in the stmtLst.

Parser::stmt is modified to return the statement number of the first
statement parsed, and this statement number is used by Parser::stmt_lst
to store the stmtLst number into PKB.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>