Test_00_SelectNoClause - add test case for selecting one synonym attribute
The newly added test case is called test_select_one_syn_attr.
The attribute for each synonym still returns the same
value as the synonym, except for call statements, where we need to
check if we are selecting (call or equivalently, call.stmt#) or
call.procName. In the case of call.procName, we have to return the
name of the procedure being called.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>