QE: Enable Modifies(proc,var) queries
This requires us to add in a dispatch table for Modifies(proc,var)
type of queries in QueryEvaluator's constructor.

For PKB, we added the following methods:
- get_all_vars_modified_by_procedure
- get_all_procedures_modifying_var

For ProcTable, we converted get_index to a const method, and added
the get_modifies(procName) method.

For VarTable, we added a get_all_procedures_modifying(var) method

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>