IntegrationTesting - Create SetWrapper class and write test cases in terms of it
Previously, we were making use of TestHelper::compare_string_sets to
compare the expected and actual output from our Integration tests.
While this is convenient, whenever there is an actual error, the line
number that shows up in the CPPUnit's assertion is that of
TestHelper::compare_string_sets and not the actual line where that
helper method was invoked.

Seeing how we needed to compare string sets, the StringWrapper template
class is created. There is a concrete instantiation of this template
class in terms of std::string. In this template class, the operator==
is overloaded for equality testing purposes in CPPUnit and operator<<
is overloaded to display an internal representation of the set inside
a SetWrapper instance, which is used when an assertion involving equality
testing of 2 SetWrapper objects fail.

We proceed to rewrite our test cases under IntegrationTesting/ using
the SetWrapper class.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>