EvalPKBDispatch, PKB - make retrieval functions more generic
For example, for a Modifies(x,y) call, assuming both x and y are
synonym types, x can be an assign, call, if, while, stmt, prog_line;
whereas y can be a variable. If we were to write separate PKB methods
for handling each of the x and y, this will require us to write 6
slightly differing methods. This is what we have been doing so far.

While this is still tractable for Modifies clause, the same cannot
be said for Parent, Parent*, Follows, Follows*, Next, Next*. Let us
take a look at the possible synonym argument types for both arguments
of Follows.

For Follows(x,y), x and y can both be one of assign, call, if, while,
stmt, prog_line. This is a total of 6 different synonym types, which
means we have to write 36 different retrieval methods. And that is
just for Follows alone.

Now, we take the approach of making the PKB retrieval functions generic.
Instead of writing slightly differing code for each concrete type,
we let the retrieval functions take in 2 additional arguments of
DesignEnt, which tells us the synonym types for argument 1 and 2 of
a given RelRef. From there, the generic function can do type specific
operations. This saves us the effort of maintaining and debugging
many slightly differing functions.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>