QE - Evaluate clauses based on synonym subgraphs
We evaluated clauses sequentially after parsing each query. However,
this is not an optimal strategy.

After some planning, we decided to evaluate PQL queries by constructing
a synonym graph. Each synonym has its own vertex in the graph, and
there exists an edge in the graph if and only if 2 distinct synonyms
are inside the same clause. Subsequently, the graph is traversed and
colored, then clauses are split into the connected component to which
they belong.

Our misinterpretation of the query evaluation process also required
us to change the intermediate representation of our results. This requires
us to replace ResultsGraph with ResultsTable and will be done in future
patches.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>