PKB - correct has_X methods
Previously, the has_X methods would return true as long as any type of
statement with the given statement number exists inside the SIMPLE
program. This does not take into account of the different statement
types (assign, call, if, while).

Now, we modify the has_X methods to make use of StmtBank::is_stmtType
to check the existence of the statement and that it is of the correct
type.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>