ResultsTable::absorb_table - Allow absorption of dead Table or NULL Table
Previously, we have an assertion that ensures that the Table* argument to
the ResultsTable::absorb_table method is alive.
This means that we will not be able to absorb Table* which are dead
or take in a NULL pointer for that matter.

Now, we allow absorption of dead Table* or NULL Table* . These simply
kill off the ResultsTable. We also return the alive status of the
ResultsTable through this method.

This change is necessary for handling such queries:

    assign a; while w;
    Select <a,w>

in the case where there may not be any while statements inside
the SIMPLE program.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>