Register Test21_Uses_N2S tests + correct PKB::uses_X_Y_smth_string_Y
Previously, we forgot to include the following line in the
IntegrationTesting/Test21_Uses_N2S.cpp file:

    CPPUNIT_TEST_SUITE_REGISTRATION(Test21_Uses_N2S)

so the tests cases in that file were not actually ran.

Now that the tests are running, we correct a faulty test case in
that file.

In addition, a bug was discovered with PKB::uses_X_Y_smth_string_Y ,
in that we have changed the structure of the VarTable with regards
to the Uses Relation.

We now correct this method to check the individual Use Sets (assign, if,
while, call, stmt, procedure) for non emptiness to indicate that
at least something is using a given variable.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>