Handle same synonym arguments for Relation clauses
Previously, Relation clauses with the same argument will be handled
by one of the ev_rr_ss_X_Y_00 methods or ev_rr_ss_X_Y_11 methods, which
will start transactions on a ResultsTable. Inside the ResultsTable
transactions are assertions to check that the two synonym arguments
are different.

For instance, in ResultsTable::syn_00_transaction, there is an
assertion at line 268 that does so.

Therefore, for Relation clauses with 2 synonym arguments that are the
same, this will result in an assertion failure.

Now, we correct this by introducing two new evaluation methods
for 2 same synonym arguments, namely:

    QueryEvaluator::ev_rr_ss_int_int_0

and

    QueryEvaluator::ev_rr_ss_int_int_1

There is no need for other types of arguments because queries with
two same synonym arguments can only be true for Next, Next*, Affects
and Affects*, which only take in statement type synonym arguments (and
these are all of integer type). For all the other types of Relation,
we just trigger an assertion failure.

With this change, the test cases introduced in commit
dcbdd0cc "Tested some Next." now pass.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>