QueryInfo - add alive/dead information for query
This is to enable future patches to kill off PQL queries at the parsing
phase when we encounter any clause which can never be true, such as:

    Uses(p,p)

where p is a synonym of type procedure.

Having the PQLParser kill such queries in future patches allows us to
skip query evaluation entirely when encountering such clauses, thereby
saving a lot of time.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>