PQLParser::eat_ref - add parsing of synonym
The PQL grammar defines ref as:
    ref = attrRef | synonym | "IDENT" | INTEGER

However, in previous patches, we did not program eat_ref to recognize
synonyms. Now, this is augmented in PQLParser::eat_ref

Since ref can only appear inside a with clause, we allow it to fail
more aggressively. In particular, it will now fail when parsing
a synonym succeeds but the synonym is:
(a) undeclared
OR
(b) declared but not of type prog_line. The reason is that all synonyms
    must have an attrName except for prog_line. Other synonyms must
    parse as attrRef (which requires an attrName) instead of synonym.

Some preliminary tests have been added to test this new functionality
and its associated error conditions.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>