Correct the attrRef_to_RefSynType function
Previously, we assumed that synonyms of all types could be in a
with clause without an attrName. It turns out that only synonyms of
type prog_line bear that privilege, whereas all other synonyms in a
with clause need to have an attrName.

We correct the attrRef_to_RefSynType function to reflect this fact.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>