WithClause - test type mismatch between left and right ref
This is done in
Test_20_PQLParser_With::test_err_parse_withclause_type_mismatch

Examples of a with clause with triggers such a type mismatch is:

    if1.stmt# = "someVariable"
    call1.procName = a1.stmt#
    5677 = var2.varName

In addition, we discovered a bug in refSynType_to_BaseType, where we
failed to take into account of the newly added REFSYN_PROGLINE_PROGLINE_NO
value in enum RefSynType. This is now fixed.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>