With Clause - test for PARSE_WITHCLAUSE_EXPECT_WITH error
This tests for errors in chains of consecutive with clause, where
a with clause is expected but either:

1. Parsing of left ref fails, eg.

    assign a; prog_line pl;
    Select a with a.stmt# = 46 and &$&23 = pl

Observe in the above query that we expect a valid Ref for the left Ref
of the second with clause but we got "&$&23" instead.

2. Parsing of '=' fails, eg.

    assign a; variable v;
    Select <v.varName,a> with 67 = a.stmt# and v.varName "h1"

Observe that there is no '=' sign after v.varName and before string "h1".

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>