WithClause - add normalize method
WithClause::normalize is a public method which normalizes a WithClause
to a standard representation. Its main purpose is to standardize our
representation of WithClause so as to simplify query evaluation.

For WithClause with both LHS and RHS Ref being concrete values, we do
nothing to them.
For WithClause with 1 AttrRef and 1 concrete value, we perform the
necessary swapping required to move the AttrRef to the LHS.
For WithClause with 2 AttrRef,
- If the RefSynType of LHS and RHS are the same, we order the 2 Refs by
  the lexicographical ordering of their synony,
- Otherwise, we order the two Refs according to the order of their
  RefSynType as seen in the enum RefSynType definition

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>